moved stuff around and deleted obsolete modules
