+++ pliers/converters/__init__.py
@@ -3,6 +3,8 @@ from pliers.transformers import Transformer, CollectionStimMixin
 from six import with_metaclass
 from tempfile import mkdtemp
 from joblib import Memory
+import importlib
+
 
 cachedir = mkdtemp()
 memory = Memory(cachedir=cachedir, verbose=0)
@@ -41,3 +43,40 @@ class Converter(with_metaclass(ABCMeta, Transformer)):
 
     def _transform(self, stim, *args, **kwargs):
         return self.convert(stim, *args, **kwargs)
+
+
+def get_converter(in_type, out_type):
+    ''' Scans the list of available Converters and returns an instantiation
+    of the first one whose input and output types match those passed in.
+    Args:
+        in_type (type): The type of input the converter must have.
+        out_type (type): The type of output the converter must have.
+    '''
+
+    # Recursively get all classes that inherit from the passed base class
+    def get_subclasses(cls):
+        subclasses = []
+        for sc in cls.__subclasses__():
+            subclasses.append(sc)
+            subclasses.extend(get_subclasses(sc))
+        return subclasses
+
+    from pliers.converters import Converter
+    base = Converter
+
+    bm = base.__module__
+    submods = importlib.import_module(bm).__all__
+    sources = ['%s.%s' % (bm, sm) for sm in submods]
+    [importlib.import_module(s) for s in sources]
+
+    transformers = get_subclasses(base)
+    for a in transformers:
+        concrete = len(a.__abstractmethods__) == 0
+        if a._input_type == in_type and a._output_type == out_type and concrete:
+            try:
+                conv = a()
+                return conv
+            except ValueError:
+                # Important for API converters
+                pass
+    return None
