+++ pliers/tests/test_transformers.py
@@ -1,0 +1,12 @@
+from pliers.transformers import get_transformer
+from pliers.extractors import Extractor
+from pliers.extractors.audio import STFTAudioExtractor
+
+
+def test_get_transformer_by_name():
+    tda = get_transformer('stFtAudioeXtrActOr', base=Extractor)
+    assert isinstance(tda, STFTAudioExtractor)
+
+
+
+
