+++ pliers/transformers.py
@@ -31,6 +31,7 @@ class Transformer(with_metaclass(ABCMeta)):
 
     def _validate(self, stim):
         if not isinstance(stim, self._input_type):
+            from pliers.converters import get_converter
             converter = get_converter(type(stim), self._input_type)
             if converter:
                 stim = converter.transform(stim)
@@ -62,42 +63,6 @@ class BatchTransformerMixin():
     def transform(self, stims, *args, **kwargs):
         return self._transform(self._validate(stims), *args, **kwargs)
 
-
-def get_converter(in_type, out_type):
-    ''' Scans the list of available Converters and returns an instantiation
-    of the first one whose input and output types match those passed in.
-    Args:
-        in_type (type): The type of input the converter must have.
-        out_type (type): The type of output the converter must have.
-    '''
-
-    # Recursively get all classes that inherit from the passed base class
-    def get_subclasses(cls):
-        subclasses = []
-        for sc in cls.__subclasses__():
-            subclasses.append(sc)
-            subclasses.extend(get_subclasses(sc))
-        return subclasses
-
-    from pliers.converters import Converter
-    base = Converter
-
-    bm = base.__module__
-    submods = importlib.import_module(bm).__all__
-    sources = ['%s.%s' % (bm, sm) for sm in submods]
-    [importlib.import_module(s) for s in sources]
-
-    transformers = get_subclasses(base)
-    for a in transformers:
-        concrete = len(a.__abstractmethods__) == 0
-        if a._input_type == in_type and a._output_type == out_type and concrete:
-            try:
-                conv = a()
-                return conv
-            except ValueError:
-                # Important for API converters
-                pass
-    return None
 
 def get_transformer(name, base=None, *args, **kwargs):
     ''' Scans list of currently available Transformer classes and returns an
